+++ fabric-loader-shim/src/main/java/net/fabricmc/loader/api/FabricLoader.java
@@ -1,12 +1,14 @@
 package net.fabricmc.loader.api;
 
 import dev.loaderbridge.fabric.runtime.BridgeFabricLoader;
+import java.io.File;
 import java.nio.file.Path;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Consumer;
 import net.fabricmc.api.EnvType;
+import net.fabricmc.loader.api.entrypoint.EntrypointContainer;
 
 /** Common Fabric Loader API surface. Unsupported methods are added as compatibility grows. */
 public interface FabricLoader {
@@ -16,9 +18,9 @@ public interface FabricLoader {
 
     <T> List<T> getEntrypoints(String key, Class<T> type);
 
-    default <T> void invokeEntrypoints(String key, Class<T> type, Consumer<? super T> invoker) {
-        getEntrypoints(key, type).forEach(invoker);
-    }
+    <T> List<EntrypointContainer<T>> getEntrypointContainers(String key, Class<T> type);
+
+    <T> void invokeEntrypoints(String key, Class<T> type, Consumer<? super T> invoker);
 
     ObjectShare getObjectShare();
 
@@ -32,9 +34,20 @@ public interface FabricLoader {
 
     EnvType getEnvironmentType();
 
+    @Deprecated
+    Object getGameInstance();
+
     Path getGameDir();
 
+    @Deprecated
+    File getGameDirectory();
+
     Path getConfigDir();
+
+    @Deprecated
+    File getConfigDirectory();
+
+    String[] getLaunchArguments(boolean sanitize);
 
     boolean isDevelopmentEnvironment();
 }