+++ fabric-loader-shim/src/main/java/net/fabricmc/loader/api/ModContainer.java
@@ -3,8 +3,10 @@ package net.fabricmc.loader.api;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
+import java.util.Collection;
 import java.util.Optional;
 import net.fabricmc.loader.api.metadata.ModMetadata;
+import net.fabricmc.loader.api.metadata.ModOrigin;
 
 public interface ModContainer {
     ModMetadata getMetadata();
@@ -12,6 +14,24 @@ public interface ModContainer {
     List<Path> getRootPaths();
 
     default Optional<Path> findPath(String file) {
-        return getRootPaths().stream().map(root -> root.resolve(file)).filter(Files::exists).findFirst();
+        return getRootPaths().stream()
+                .map(root -> root.resolve(file.replace("/", root.getFileSystem().getSeparator())))
+                .filter(Files::exists)
+                .findFirst();
     }
+
+    ModOrigin getOrigin();
+
+    Optional<ModContainer> getContainingMod();
+
+    Collection<ModContainer> getContainedMods();
+
+    @Deprecated
+    default Path getRoot() { return getRootPath(); }
+
+    @Deprecated
+    Path getRootPath();
+
+    @Deprecated
+    Path getPath(String file);
 }