+++ fabric-loader-shim/src/main/java/net/fabricmc/loader/api/ObjectShare.java
@@ -1,7 +1,11 @@
 package net.fabricmc.loader.api;
 
+import java.util.function.BiConsumer;
+
 public interface ObjectShare {
     Object get(String key);
+
+    void whenAvailable(String key, BiConsumer<String, Object> consumer);
 
     Object put(String key, Object value);