+++ fabric-loader-shim/src/main/java/net/fabricmc/loader/api/VersionParsingException.java
@@ -1,0 +1,9 @@
+package net.fabricmc.loader.api;
+
+@SuppressWarnings({"deprecation", "serial"})
+public class VersionParsingException extends net.fabricmc.loader.util.version.VersionParsingException {
+    public VersionParsingException() {}
+    public VersionParsingException(Throwable cause) { super(cause); }
+    public VersionParsingException(String message) { super(message); }
+    public VersionParsingException(String message, Throwable cause) { super(message, cause); }
+}