+++ fabric-loader-shim/src/main/java/net/fabricmc/loader/api/metadata/ModOrigin.java
@@ -1,0 +1,17 @@
+package net.fabricmc.loader.api.metadata;
+
+import java.nio.file.Path;
+import java.util.List;
+
+public interface ModOrigin {
+    Kind getKind();
+    List<Path> getPaths();
+    String getParentModId();
+    String getParentSubLocation();
+
+    enum Kind {
+        PATH,
+        NESTED,
+        UNKNOWN
+    }
+}