+++ fabric-loader-shim/src/main/java/net/fabricmc/loader/api/metadata/version/VersionInterval.java
@@ -1,0 +1,11 @@
+package net.fabricmc.loader.api.metadata.version;
+
+import net.fabricmc.loader.api.Version;
+
+public interface VersionInterval {
+    boolean isSemantic();
+    Version getMin();
+    boolean isMinInclusive();
+    Version getMax();
+    boolean isMaxInclusive();
+}