+++ fabric-loader-shim/src/main/java/net/fabricmc/loader/api/metadata/version/VersionPredicate.java
@@ -1,0 +1,12 @@
+package net.fabricmc.loader.api.metadata.version;
+
+import java.util.Collection;
+import java.util.function.Predicate;
+import net.fabricmc.loader.api.Version;
+
+public interface VersionPredicate extends Predicate<Version> {
+    Collection<? extends PredicateTerm> getTerms();
+    VersionInterval getInterval();
+
+    interface PredicateTerm extends Predicate<Version> {}
+}